Refactored folder structure. All under source. Conditional compiling for makefile based pal implementations
