+++ src/common/hall-switch.cpp
