+++ src/common/hall-switch.h
